requirements.txt
langchain
openai
tiktoken
pinecone-client
pinecone-text
streamlit
networkx